+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - feat: add readfrom json tag to support reverse edges
   [#49](https://github.com/hypermodeinc/modusDB/pull/49)
 
+- chore: Refactoring package management #51 [#51](https://github.com/hypermodeinc/modusDB/pull/51)
+
 ## 2025-01-02 - Version 0.1.0
 
 Baseline for the changelog.
